fix(agent): enforce `IPage<T>` type
